v1.2.1 New feature : * A confirmation is now required to delete an item of the character sheet Fixes : * Character description won't saves in FVTT 13 * Item directory was not functionnal inFVTT 13 * Dice so nices was not triggered on Luck roll and Action rolls.

+++ sysfeerie/modules/sheets/actor-sheet.js
@@ -168,9 +168,11 @@ export class SFActorSheet extends ActorSheet {
 
 		// Delete item/status
 		html.find('.item-delete').click(ev => {
-			const li = $(ev.currentTarget).parents(".item");
-			this.actor.deleteEmbeddedDocuments("Item",[li.data("itemId")]);
-			li.slideUp(200, () => this.render(false));
+			if(window.confirm(game.i18n.localize("SYSFEERIE.Item.DeleteConfirm")))  {
+				const li = $(ev.currentTarget).parents(".item");
+				this.actor.deleteEmbeddedDocuments("Item",[li.data("itemId")]);
+				li.slideUp(200, () => this.render(false));
+			}
 			ev.stopPropagation();
 		});
